# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RejeXion API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  /** @minLength 8 */
  password: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface Profile {
  id: string;
  email: string;
  name: string;
  totalPoints: number;
  entryCount: number;
  growthNoteCount: number;
  createdAt: string;
}

export type PromptTemplateCategory =
  (typeof PromptTemplateCategory)[keyof typeof PromptTemplateCategory];

export const PromptTemplateCategory = {
  academic: "academic",
  career: "career",
} as const;

export interface PromptTemplate {
  id: string;
  category: PromptTemplateCategory;
  promptText: string;
  orderIndex: number;
}

export interface UpdateEntryBody {
  title?: string;
  rejectionDate?: string;
  description?: string;
}

export type CreateEntryBodyCategory =
  (typeof CreateEntryBodyCategory)[keyof typeof CreateEntryBodyCategory];

export const CreateEntryBodyCategory = {
  academic: "academic",
  career: "career",
} as const;

export interface EntryResponseInput {
  promptId: string;
  responseText: string;
}

export interface CreateEntryBody {
  category: CreateEntryBodyCategory;
  title: string;
  rejectionDate: string;
  description: string;
  responses: EntryResponseInput[];
}

export type EntryCategory = (typeof EntryCategory)[keyof typeof EntryCategory];

export const EntryCategory = {
  academic: "academic",
  career: "career",
} as const;

export interface Entry {
  id: string;
  category: EntryCategory;
  title: string;
  rejectionDate: string;
  description: string;
  growthNoteCount: number;
  createdAt: string;
}

export interface EntryResponseItem {
  id: string;
  promptId: string;
  promptText: string;
  responseText: string;
}

export type EntryDetailCategory =
  (typeof EntryDetailCategory)[keyof typeof EntryDetailCategory];

export const EntryDetailCategory = {
  academic: "academic",
  career: "career",
} as const;

export interface GrowthNote {
  id: string;
  entryId: string;
  content: string;
  createdAt: string;
}

export interface EntryDetail {
  id: string;
  category: EntryDetailCategory;
  title: string;
  rejectionDate: string;
  description: string;
  responses: EntryResponseItem[];
  growthNotes: GrowthNote[];
  createdAt: string;
}

export interface CreateGrowthNoteBody {
  content: string;
}

export interface PointsSummary {
  totalPoints: number;
  entryPoints: number;
  growthNotePoints: number;
  entryCount: number;
  growthNoteCount: number;
}

export interface StatsSummary {
  totalEntries: number;
  academicEntries: number;
  careerEntries: number;
  totalGrowthNotes: number;
  totalPoints: number;
  recentEntries: Entry[];
}

export type GetPromptsParams = {
  category: GetPromptsCategory;
};

export type GetPromptsCategory =
  (typeof GetPromptsCategory)[keyof typeof GetPromptsCategory];

export const GetPromptsCategory = {
  academic: "academic",
  career: "career",
} as const;

export type ListEntriesParams = {
  category?: ListEntriesCategory;
};

export type ListEntriesCategory =
  (typeof ListEntriesCategory)[keyof typeof ListEntriesCategory];

export const ListEntriesCategory = {
  academic: "academic",
  career: "career",
  all: "all",
} as const;