# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RejeXion API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List user entries
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [academic, career, all]
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a new rejection entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryBody"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get a specific entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry's basic details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryBody"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Entry deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries/{id}/export:
    get:
      operationId: exportEntry
      tags: [entries]
      summary: Export entry as plain text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plain text export
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - token
        - password
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
...
    UpdateEntryBody:
      type: object
      properties:
        title:
          type: string
        rejectionDate:
          type: string
          format: date
        description:
          type: string
    CreateEntryBody:
      type: object
      properties:
        category:
          type: string
          enum: [academic, career]
        title:
          type: string
        rejectionDate:
          type: string
          format: date
        description:
          type: string
        responses:
          type: array
          items:
            $ref: "#/components/schemas/EntryResponseInput"
      required:
        - category
        - title
        - rejectionDate
        - description
        - responses
    EntryResponseInput:
      type: object
      properties:
        promptId:
          type: string
        responseText:
          type: string
      required:
        - promptId
        - responseText
    Entry:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
          enum: [academic, career]
        title:
          type: string
        rejectionDate:
          type: string
          format: date
        description:
          type: string
        growthNoteCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - title
        - rejectionDate
        - description
        - growthNoteCount
        - createdAt
    EntryResponseItem:
      type: object
      properties:
        id:
          type: string
        promptId:
          type: string
        promptText:
          type: string
        responseText:
          type: string
      required:
        - id
        - promptId
        - promptText
        - responseText
    EntryDetail:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
          enum: [academic, career]
        title:
          type: string
        rejectionDate:
          type: string
          format: date
        description:
          type: string
        responses:
          type: array
          items:
            $ref: "#/components/schemas/EntryResponseItem"
        growthNotes:
          type: array
          items:
            $ref: "#/components/schemas/GrowthNote"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - title
        - rejectionDate
        - description
        - responses
        - growthNotes
        - createdAt
    GrowthNote:
      type: object
      properties:
        id:
          type: string
        entryId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - entryId
        - content
        - createdAt
...