# lib/api-zod/src/generated/types/updateEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RejeXion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEntryBody {
  title?: string;
  rejectionDate?: Date;
  description?: string;
}